requirements.txt
PyQt6
pandas
openpyxl
requests
python-dotenv
packaging